factor filterHistoryImages into a pure helper with direct tests
