+++ extensions/copilot/src/platform/endpoint/node/chatEndpoint.ts
@@ -34,6 +34,7 @@ import { IDomainService } from '../common/domainService';
 import { CustomModel, IChatModelInformation, ModelSupportedEndpoint } from '../common/endpointProvider';
 import { createMessagesRequestBody, processResponseFromMessagesEndpoint } from './messagesApi';
 import { createResponsesRequestBody, getResponsesApiCompactionThreshold, processResponseFromChatEndpoint } from './responsesApi';
+import { filterHistoryImages } from './imageLimits';
 
 /**
  * The default processor for the stream format from CAPI
@@ -290,7 +291,7 @@ export class ChatEndpoint implements IChatEndpoint {
 		// Determine per-model image limit for APIs with known restrictions
 		const imageLimit = this.getImageLimit();
 		if (imageLimit !== undefined) {
-			options = { ...options, messages: this.validateAndFilterImages(options.messages, imageLimit) };
+			options = { ...options, messages: filterHistoryImages(options.messages, imageLimit) };
 		}
 
 		if (this.useResponsesApi) {
@@ -318,115 +319,6 @@ export class ChatEndpoint implements IChatEndpoint {
 			return this.maxPromptImages ?? 10;
 		}
 		return undefined;
-	}
-
-	/**
-	 * Silently drops the oldest images from history when the total number of images
-	 * in the conversation exceeds `maxImages`. Images belonging to the current turn
-	 * (the last user message and anything after it, e.g. recent tool results) are
-	 * always preserved — if the current turn alone exceeds the limit we let the
-	 * request through and rely on the model to surface an error.
-	 *
-	 * @returns A (possibly filtered) copy of messages. The original array is never mutated.
-	 */
-	private validateAndFilterImages(messages: Raw.ChatMessage[], maxImages: number): Raw.ChatMessage[] {
-		// Anchor the current turn at the last user message; anything at or after this
-		// index is treated as "current turn" and its images are never filtered.
-		let lastUserIdx = -1;
-		for (let i = messages.length - 1; i >= 0; i--) {
-			if (messages[i].role === Raw.ChatRole.User) {
-				lastUserIdx = i;
-				break;
-			}
-		}
-
-		// Corner case: no user message at all (e.g. system-only history). Treat the
-		// last message as the current turn so we still filter earlier images.
-		if (lastUserIdx === -1 && messages.length > 0) {
-			lastUserIdx = messages.length - 1;
-		}
-
-		// Count images in the current turn (the last user message and anything after it).
-		let currentTurnImages = 0;
-		for (let i = Math.max(lastUserIdx, 0); i < messages.length; i++) {
-			const content = messages[i].content;
-			if (!Array.isArray(content)) {
-				continue;
-			}
-			for (const part of content) {
-				if (part.type === Raw.ChatCompletionContentPartKind.Image) {
-					currentTurnImages++;
-				}
-			}
-		}
-
-		// Count total images across all messages
-		let totalImages = 0;
-		for (const message of messages) {
-			if (Array.isArray(message.content)) {
-				for (const part of message.content) {
-					if (part.type === Raw.ChatCompletionContentPartKind.Image) {
-						totalImages++;
-					}
-				}
-			}
-		}
-
-		// No filtering needed if total is within the limit
-		if (totalImages <= maxImages) {
-			return messages;
-		}
-
-		// Walk backward through history (before the current turn), keeping the
-		// most recent images and replacing the oldest with placeholders. Budget
-		// can go negative if the current turn alone exceeds the limit — in that
-		// case we drop all history images and let the model handle it.
-		let historyBudget = maxImages - currentTurnImages;
-
-		// Collect keep/drop decisions by walking backward through history
-		const historyImageDecisions = new Map<string, boolean>(); // "msgIdx:partIdx" -> keep
-		for (let i = lastUserIdx - 1; i >= 0; i--) {
-			if (!Array.isArray(messages[i].content)) {
-				continue;
-			}
-			for (let j = messages[i].content.length - 1; j >= 0; j--) {
-				if (messages[i].content[j].type === Raw.ChatCompletionContentPartKind.Image) {
-					const key = `${i}:${j}`;
-					if (historyBudget > 0) {
-						historyImageDecisions.set(key, true);
-						historyBudget--;
-					} else {
-						historyImageDecisions.set(key, false);
-					}
-				}
-			}
-		}
-
-		// Build filtered messages, replacing dropped images with text placeholders
-		return messages.map((message, msgIdx) => {
-			if (msgIdx >= lastUserIdx) {
-				return message;
-			}
-			if (!Array.isArray(message.content)) {
-				return message;
-			}
-			if (!message.content.some(p => p.type === Raw.ChatCompletionContentPartKind.Image)) {
-				return message;
-			}
-			return {
-				...message,
-				content: message.content.map((part, partIdx) => {
-					if (part.type !== Raw.ChatCompletionContentPartKind.Image) {
-						return part;
-					}
-					if (historyImageDecisions.get(`${msgIdx}:${partIdx}`)) {
-						return part;
-					}
-					// Model-facing placeholder; intentionally not localized.
-					return { type: Raw.ChatCompletionContentPartKind.Text, text: '[Image omitted from conversation history due to model limit.]' };
-				})
-			};
-		});
 	}
 
 	protected getCompletionsCallback(): RawMessageConversionCallback | undefined {
